Update default value in CPSubsystemConfig
missingCPMemberAutoRemovalSeconds is set to 4 hours by default.

+++ hazelcast/src/main/java/com/hazelcast/internal/dynamicconfig/DynamicCPSubsystemConfig.java
@@ -53,7 +53,7 @@ class DynamicCPSubsystemConfig extends CPSubsystemConfig {
     }
 
     @Override
-    public CPSubsystemConfig setMissingCPMemberAutoRemovalSeconds(long missingCPMemberAutoRemovalSeconds) {
+    public CPSubsystemConfig setMissingCPMemberAutoRemovalSeconds(int missingCPMemberAutoRemovalSeconds) {
         throw new UnsupportedOperationException();
     }
